backend/requirements.txt
# Django
Django>=5.1,<5.2
djangorestframework>=3.15
django-cors-headers>=4.4
django-allauth>=65.0
django-filter>=24.0
django-ratelimit>=4.1
dj-database-url>=2.2
django-storages[boto3]>=1.14

# Channels (WebSocket)
channels>=4.1
channels-redis>=4.2
daphne>=4.1

# Database
psycopg[binary]>=3.2
pgvector>=0.3

# Task Queue
celery[redis]>=5.4
django-celery-beat>=2.6

# AWS
boto3>=1.35
django-ses>=4.1

# AI/ML
google-generativeai>=0.8
anthropic>=0.32
openai>=1.40
tiktoken>=0.7

# File Processing
pypdf>=4.3
python-docx>=1.1
python-pptx>=0.6
trafilatura>=1.12
beautifulsoup4>=4.12

# Utilities
python-dotenv>=1.0
Pillow>=10.4
gunicorn>=22.0
whitenoise>=6.7
bleach>=6.1
python-slugify>=8.0
pydantic>=2.8
structlog>=24.4

# Development
pytest>=8.3
pytest-django>=4.8
pytest-cov>=5.0
ruff>=0.5
mypy>=1.11
django-debug-toolbar>=4.4
factory-boy>=3.3
